Ok so, I made the While loop work finally. damn. Very hard. But ok. So, Got it to pipe the value of difficulty to the function that generates the ship. So we can have two ships. I have yet to get it to save ANY of that data, but next step I think is saving the data for the ship that you choose.
Maybe this project won't be as big as I want it to be, and it'll be a smaller littler one that has simple
skirmishes, but I'd love to create a static map, maybe random gen. shrug.jpg. That the player can spawn in and do
things in. We'll see how I progress, I could use some assistance. A friend in this.